[PATCH] bio_add_page()
This is bio_add_page(), 100% identical to the version I sent out for
comments earlier this week.  With the previous queue restriction patch,
this guarentees that we can always add a page worth of data to the bio.
bio_add_page() returns 0 on success, and 1 on failure.  Either the page
is added completely, or the attempt is aborted.

bio_add_page() uses the normal queue restrictions to determine whether
we an add the page or not. if a queue has further restrictions, it can
define a q->merge_bvec_fn() to further impose limits.

Patch also includes changes to ll_rw_kio(), if for nothing else to
demonstrate how to use this piece of instrastructure.

BKrev: 3d88a2feRZ8AX4lvx10EgGjEQGFekA