[PATCH] tighter locking in pdflush
Had a weird oops from Bill Irwin - the pdflush_list was corrupt.

The only thing I can think of is that something sprayed out a wakeup
when it shouldn't.  So tighten things up against that, and add some
printks to catch it if it happens again.

BKrev: 3d928b80zw6UFlGpNPzg1cPf_69IOA