[PATCH] per-node kswapd instances
Patch from David Hansen.

Start one kswapd instance for each NUMA node.  That kswapd instance
only works against the pages which are local to that node.

We need to bind that kswapd to that node's CPU set, but the
infrastructure for this is not yet in place.

BKrev: 3d978ac2MH9qcfniocB4I32JDgpE-Q