Merge http://jdike.stearns.org:5000/updates-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/09/30 00:02:33-04:00 jdike
main.o needed to be added to the vmlinux dependencies so it would build.

2002/09/29 22:43:05-04:00 jdike
Moved the linker script from vmlinux.lds.S, which will be empty, to
uml.ld.S.

2002/09/28 22:28:25-04:00 jdike
Updated to build with the 2.5.39 kbuild.

2002/09/28 22:23:28-04:00 jdike
Took the 2.5.39 Makefile changes.

2002/09/23 21:11:43-04:00 jdike
Bumped EXTRAVERSION for the 2.4 fixes and highmem support.

2002/09/23 14:18:45-04:00 jdike
One last fix to the ubd driver, allowing UML to boot.

2002/09/23 13:57:10-04:00 jdike
Trivial fix to the ubd driver.

2002/09/23 13:34:18-04:00 jdike
Cleaned up arch/um/Makefile and updated the ubd driver.

2002/09/23 12:52:56-04:00 jdike
UML updates to allow it to build and run as 2.5.38.

BKrev: 3d989002ayc7jsFekiiacjMixzzBYw