[PATCH] pcmcia resource allocation fix
The patch below is a forward-port from 2.4 of a fix that went in to
the 2.4.x PCMCIA code some time back.  It makes sure that that we
request I/O and memory regions from the correct resource (the parent
of the PCMCIA bridge chip, for PCMCIA bridges connected to a PCI bus)
rather than always requesting them from the top-level ioport_resource
or iomem_resource.

BKrev: 3d99c2b69IyUR7QbkfU5Iau0taC-7Q