Merge mulgrave.(none):/home/jejb/BK/scsi-reject-zero-2.5 into mulgrave.(none):/home/jejb/BK/scsi-for-linus-2.5
2002/10/02 14:18:16-04:00 jejb
[SCSI] remove comment that every host is expected to be able to queue at least one command

2002/10/01 18:14:18-04:00 jejb
[SCSI] remove debugging from zero depth queue handling

2002/10/01 18:01:59-04:00 jejb
[PATCH SCSI] make BUSY status stall the device for a while

2002/10/01 16:44:42-04:00 jejb
[PATCH] remove mid-layer assumption that devices must be able to queue at least one command

This allows the request_fn() to recover properly from either host_blocked
or device_blocked at zero command depth.

Also adds the facility for queuecommand() to tell us whether it wants
the host or only the device blocked

2002/10/01 10:28:11-04:00 jejb
[PATCH] first cut at fixing unable to requeue with no outstanding
 commands

The attached represents an attempt to break the scsi mid-layer of the
assumption that any device can queue at least one command.

What essentially happens if the host rejects a command with no other
outstanding commands, it does a very crude countdown (basically counts the
number of cycles through the scsi request function) until the device gets
enabled again when the count reaches zero.  I think the iteration in the
request function is better than a fixed timer because it makes the system more
responsive to I/O pressure (and also, it's easier to code).

I've tested this by making a SCSI driver artificially reject commands with
none outstanding (and run it on my root device).  A value of seven seems to
cause a delay of between half and five seconds before the host starts up again
(depending on the I/O load).

If this approach looks acceptable, I plan the following enhancements

1. Make device_busy count down in the same fashion
2. give ->queuecommand() a two value return (one for blocking the entire host
and another for just blocking the device).
3. Make the countdown tuneable from the host template.

BKrev: 3d9d950054znrcCzpofDZFji_zHUIQ