[PATCH] fix reclaim for higher-order allocations
The page reclaim logic will bail out if all zones are at pages_high.
But if the caller is requesting a higher-order allocation we need to go
on and free more memory anyway.  That's the only way we have of
addressing buddy fragmentation.

BKrev: 3d9f22d15KODDhscsHeAiYBp2Rs0HQ