[PATCH] NCR5380 port to 2.5 first pass
There is still more work to do, the driver sucks in 2.4 and 2.5 but 2.5 has a
lot more of what is needed to make it work nicely. Basically NCR5380_main
probably has to become a thread in the next generation of the code.

This however seems to get it up and crawling

BKrev: 3da0bfd0MtLGNTDJr-v34b0akBPSFQ