[PATCH] NUMA-Q fixes
This patch basically just renames usages of CONFIG_MULTIQUAD to
CONFIG_X86_NUMAQ. The original multiquad option covered a variety
of sins, and just made a mess (my fault). After extensive community
discussion, this is the naming scheme everyone agreed to - the
CONFIG_X86_NUMAQ option already exists, we're just renaming usages
of MULTIQUAD to it, and removing the MULTIQUAD option.

Stuff that's really to do with clustered APIC mode is put under
CONFIG_CLUSTERED_APIC instead, as it's not specific to the NUMA-Q.

Note that the definitions of clustered_apic_mode are still duplicated,
so the changes appear twice.  I'm not making it worse, just haven't
fixed this one yet, will do that next.

Patch was written by Matt Dobson.

BKrev: 3da31715MZ20Cd_VQOkchrRmK4FJYA