Merge osdl.org:/home/mochel/src/kernel/devel/linux-2.5-virgin into osdl.org:/home/mochel/src/kernel/devel/linux-2.5-core
2002/10/08 17:32:56-07:00 mochel
USB: call device_unregister() instead of put_device() when removing devices.

2002/10/08 17:32:17-07:00 mochel
IDE: call device_unregister() instead of put_device() in ide-disk->cleanup().

2002/10/08 17:23:56-07:00 mochel
driver model: check return of get_device() when creating a driverfs file.

2002/10/08 17:23:05-07:00 mochel
driver model: and present field to struct device and implement device_unregister().

device_unregister() is intended to be the complement of device_register(), and assumes
most of the functionality of the current put_device(). It should be called by the bus
driver when a device is physically removed from the system.

It should _not_ be called from a driver's remove() method, as that remove() method is called
via device_detach() in device_unregister().

dev->present is used to flag the physical presence of the device. It is set when the device
is registered, and cleared when unregistered. get_device() checks this flag, and returns
a NULL pointer if cleared. This prevents anyone from obtaining a reference to a device
that has been unregistered (removed), but not yet been freed (e.g. if someone else is
holding a reference to it).

put_device() BUG()s if dev->present is set. A device should be unregistered before it is
freed. This will catch people doing it in the wrong order.

BKrev: 3da4452cLm4jLalImvg18jHXCPTRSw