[PATCH] Fix xxx_get_biosgeometry --- avoid useless 64-bit division.
From Peter Chubb

The bios geometry is almost useless, except for fdisk to try to write
an MSDOS partition table that is vaguely compatible with one written by
other operating systems.

If the size of disc will overflow a ten-bit cylinder number, then all
bets are off anyway.  So fake it by casting the true disc capacity to a
smaller type (than u64), so that we avoid 64-bit division on 32-bit
platforms.  If the disc is small enough that the number of cylinders is
correct, then this has no effect; otherwise, the number-of-cylinders we
report is bogus, but you can't use an MSDOS-format partition table on
such a drive anyway --- use the EFI GPT or the LDM partitioning, which
use 64-bit offsets internally.

BKrev: 3da44d550OGWeXHoBUWP7fcSBxu1Bg