[PATCH] initcalls for ATM
Switch over ATM code to initcalls and reorder the makefile so
that link order inside atm is the same.  I've also cleaned up
the makefile a bit while at it.

I didn't fix the existing compilation problems in the drivers (cli &
friends) and the broken le/be firmware selection for the fore200e cards
(kbuild breakage) though.

BKrev: 3da46441urwHTCGZaFLIYDEikmNqcw